Add files via upload
Google News scraper with simple Tkinter GUI. Retrieves the site on which the article is hosted on, when it was published, and relevant authors for user-inputted keywords. Input is handled via Tkinter GUI and output is printed to the console as well as saved in a CSV for import to Excel or other uses.